refactor(cli/doctor): split checks.py + app.py under the 500-LOC module-health gate
checks.py (616 gate-LOC on main) breaches the module-health ratchet; app.py was
516 at #3310 (now 488 after intervening PRs, but the 5-way checks import would
push it back over). #3310 grew both past the cap and the CI gate's failure was
ignored on merge. Split both by concern so every doctor module sits well under
500 LOC and the next PR touching them does not re-trigger the gate.

checks.py -> five concern modules (function bodies moved verbatim):
  - checks_environment.py  clone/install/venv hygiene (9 probes)
  - checks_runtime.py      running-process health (3 probes)
  - checks_mcp.py          MCP/connector wiring (4 probes)
  - checks_session.py      Slack/session identity (3 probes)
  - checks_loop.py         loop/scheduling staleness (4 probes)

app.py -> DoctorService + IntrospectionHelpers (install management) moved to
service.py; app.py keeps the Typer group and the run_doctor_checks orchestrator.

app.py's __all__ re-export surface is byte-identical, so nothing outside the
doctor package changes. Consumers repointed: the orchestrator's imports, the
direct test imports of teatree.cli.doctor.checks._check_*, the string
mock.patch targets that named the old module (consistency_findings, typer.echo,
app._find_host_project_root, app.distribution), and four docstring cross-refs
(app.py, statusline.py, and the package __init__ facade, which now names the
service/checks_* modules instead of the removed checks submodule).

Behavior-preserving move only, no function body changed (verified byte-for-byte
via ast.get_source_segment against HEAD).

Gate-LOC after split: app 198, service 318, checks_environment 246,
checks_mcp 132, checks_loop 105, checks_session 98, checks_runtime 54.

Open questions & assumptions:
- (assumed) checks.py deleted rather than kept as a backward-compat re-export
  shim -- the issue's plan says to update the tests that import
  teatree.cli.doctor.checks._check_* directly, i.e. remove the module.
- (assumed) DoctorService/IntrospectionHelpers carved into service.py to give
  app.py durable headroom and honor the ticket's 'split app.py' scope, even
  though app.py alone is now under the cap.
- (decided-by-evidence) 7 env-dependent failures in test_doctor_check_command.py
  (a real shadowing t3 on the sandbox PATH) fail identically on origin/main and
  are pre-existing, not introduced by this change.
