+++ src/teatree/cli/doctor/__init__.py
@@ -1,9 +1,11 @@
 """``t3 doctor`` — smoke-test hooks, imports, and services (thin package facade).
 
-The ``doctor`` Typer group, the :class:`DoctorService` / :class:`IntrospectionHelpers`
-services, and the individual ``_check_*`` probes live in :mod:`teatree.cli.doctor.app`;
-the larger check bodies and plugin-registration repair live in the ``checks`` and
-``plugin_repair`` submodules. This ``__init__`` only re-exports the public surface so
+The ``doctor`` Typer group + the ``check`` orchestrator live in
+:mod:`teatree.cli.doctor.app`; the :class:`DoctorService` / :class:`IntrospectionHelpers`
+services live in :mod:`teatree.cli.doctor.service`; the ``_check_*`` probes live in the
+``checks_environment`` / ``checks_runtime`` / ``checks_mcp`` / ``checks_session`` /
+``checks_loop`` / ``checks_availability`` modules; plugin-registration repair lives in
+``plugin_repair``. This ``__init__`` only re-exports the public surface so
 ``from teatree.cli.doctor import DoctorService`` (and siblings) is unchanged.
 """
